:truck: Extract operator utils
