+++ @navikt/core/react/src/data/token-filter/helpers/generate-autocomplete-options.ts
@@ -5,201 +5,23 @@ import type {
   QueryFilterOperator,
 } from "../TokenFilter.types";
 import { createGroups } from "./grouping";
-import { type ParsedText, QUERY_OPERATORS } from "./parse-query-text";
+import { QUERY_OPERATORS } from "./operators";
+import { type ParsedText } from "./parse-query-text";
 import { OPERATOR_LABELS, buildQueryString } from "./query-builder";
 import { matchesFilterText } from "./text-matching";
 
 /**
- * Returns the valid operators for a given property.
- * Currently returns all operators for all properties.
- *
- * TODO: Implement per-property operator configuration.
- * This will allow properties to restrict which operators are valid.
- * For example, a boolean property might only support "=" and "!=",
- * while a text property supports ":", "!:", "^", etc.
- *
- * @returns Array of valid operators for the property
- */
-function getValidOperatorsForProperty(
-  /* TODO: Will be implemented when each property can configure operator per instance */
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  _property: ParsedProperty,
-): QueryFilterOperator[] {
-  return QUERY_OPERATORS;
-}
-
-function filterOperatorsByPrefix(
-  operators: QueryFilterOperator[],
-  prefix: string,
-): QueryFilterOperator[] {
-  if (!prefix) {
-    return operators;
-  }
-
-  return operators.filter((operator) => operator.startsWith(prefix));
-}
-
-function generatePropertySuggestions(
-  filteringProperties: ParsedProperty[] = [],
-  filterText = "",
-): OptionGroup<ParsedProperty>[] {
-  const filteredProperties = filteringProperties.filter((property) => {
-    if (!property) {
-      return false;
-    }
-
-    return matchesFilterText(
-      [
-        property.propertyLabel,
-        property.groupValuesLabel,
-        property.propertyGroup,
-      ].filter(Boolean),
-      filterText,
-    );
-  });
-
-  return createGroups(
-    filteredProperties,
-    (property) => property.propertyGroup,
-    "Properties",
-  );
-}
-
-function generateOperatorSuggestions(
-  property: ParsedProperty,
-  operatorPrefix = "",
-  filterText = "",
-): OptionGroup<AutoCompleteOption>[] {
-  const operators = filterOperatorsByPrefix(
-    getValidOperatorsForProperty(property),
-    operatorPrefix,
-  ).filter((operator) =>
-    matchesFilterText(
-      [operator, OPERATOR_LABELS[operator] ?? ""].filter(Boolean),
-      filterText,
-    ),
-  );
-
-  if (operators.length === 0) {
-    return [];
-  }
-
-  return [
-    {
-      label: "Operators",
-      options: operators.map((operator) => ({
-        value: buildQueryString(property.propertyLabel, operator, ""),
-        label: buildQueryString(property.propertyLabel, operator, ""),
-        description: OPERATOR_LABELS[operator] ?? "",
-      })),
-    },
-  ];
-}
-
-/**
- * Creates value suggestions for autocomplete.
- * When scopedProperty is provided, only shows values for that property (single group).
- * When scopedProperty is omitted, searches across all properties (multiple groups).
- */
-function createValueSuggestions(
-  filteringOptions: ParsedOption[] = [],
-  operator: QueryFilterOperator,
-  filterText = "",
-  scopedProperty?: ParsedProperty,
-): OptionGroup<AutoCompleteOption>[] {
-  const groups: Record<string, OptionGroup<AutoCompleteOption>> = {};
-
-  for (const option of filteringOptions) {
-    if (!option?.property) {
-      continue;
-    }
-
-    // If scoped to a property, filter to only that property's options
-    if (scopedProperty && option.property !== scopedProperty) {
-      continue;
-    }
-
-    // Build search fields
-    const searchFields = [
-      option.label,
-      ...(option.tags ?? []),
-      ...(option.filteringTags ?? []),
-    ];
-
-    // For free-text search (no scoped property), also search property label
-    if (!scopedProperty) {
-      searchFields.push(option.property.propertyLabel);
-    }
-
-    const matches = matchesFilterText(searchFields.filter(Boolean), filterText);
-
-    if (!matches) {
-      continue;
-    }
-
-    const groupLabel = option.property.groupValuesLabel || "Values";
-
-    if (!groups[groupLabel]) {
-      groups[groupLabel] = {
-        label: groupLabel,
-        options: [],
-      };
-    }
-
-    groups[groupLabel].options.push({
-      value: buildQueryString(
-        option.property.propertyLabel,
-        operator,
-        option.value,
-      ),
-      label: option.label,
-      tags: option.tags,
-      filteringTags: option.filteringTags,
-    });
-  }
-
-  return Object.values(groups).filter((group) => group.options.length > 0);
-}
-
-function generateValueSuggestions(
-  filteringOptions: ParsedOption[] = [],
-  property: ParsedProperty,
-  operator: QueryFilterOperator,
-  filterText = "",
-): OptionGroup<AutoCompleteOption>[] {
-  return createValueSuggestions(
-    filteringOptions,
-    operator,
-    filterText,
-    property,
-  );
-}
-
-function generateFreeTextValueSuggestions(
-  filteringOptions: ParsedOption[] = [],
-  operator: QueryFilterOperator,
-  filterText = "",
-): OptionGroup<AutoCompleteOption>[] {
-  return createValueSuggestions(filteringOptions, operator, filterText);
-}
-
-/**
- * Generates autocomplete suggestions based on the current query state.
+ * Generates "options" to be used as autosuggest-ottion based on the current query state.
  *
  * The query parser recognizes three states:
- * - "property": User has selected/matched a property and operator (e.g., "Status = active")
- * - "operator": User has matched a property but is typing the operator (e.g., "Status !")
- * - "free-text": User is typing freely without a property match (e.g., "act" or "!:test")
+ * - "property": User has selected/matched a property and operator ("Status = active")
+ * - "operator": User has matched a property but is typing the operator ("Status" or "Status !")
+ * - "free-text": User is typing freely without a property match (e.g., "act" or "!: test")
  *
- * Returns an object with:
+ * @returns
  * - value: The canonical query string representation for the current state.
  *   Used by the UI to determine cursor position and input replacement.
  * - options: Grouped suggestions to display (properties, operators, or values).
- *
- * @param queryState - Parsed query state from parseQueryText
- * @param filteringProperties - Available properties for filtering (defaults to empty array)
- * @param filteringOptions - Available values for filtering (defaults to empty array)
- * @returns Object containing the query value and suggestion groups
  */
 function generateAutoCompleteOptions(
   queryState: ParsedText,
@@ -212,16 +34,16 @@ function generateAutoCompleteOptions(
 
     return {
       value: queryState.value,
-      options: generateValueSuggestions(
+      options: createValueSuggestions(
         filteringOptions,
-        queryState.property,
         queryState.operator,
         filterText,
+        queryState.property,
       ),
     };
   }
 
-  // State: Property matched, but operator is incomplete
+  /* State: Property matched, but operator is incomplete */
   if (queryState.step === "operator") {
     const operators = filterOperatorsByPrefix(
       getValidOperatorsForProperty(queryState.property),
@@ -258,10 +80,10 @@ function generateAutoCompleteOptions(
     };
   }
 
-  // State: Free-text mode (no property match)
-
-  // Edge case: Input starts with operator but has no value yet (e.g., user typed just "!=")
-  // Don't show suggestions - wait for them to type a value
+  /*
+   * Edge case: Input starts with operator but has no value yet (user typed just "!=")
+   * Wait for value before showing suggestions
+   */
   if (!queryState.value && queryState.operator) {
     return {
       value: "",
@@ -269,7 +91,7 @@ function generateAutoCompleteOptions(
     };
   }
 
-  // Empty input: Show all properties
+  /* Empty input: Show all properties */
   if (!queryState.value) {
     return {
       value: "",
@@ -277,16 +99,186 @@ function generateAutoCompleteOptions(
     };
   }
 
-  // Free-text search: Show matching values across all properties
-  // Use the detected operator if input started with one (e.g., "!=test"), otherwise default to "="
+  /*
+   * Free-text search: Show matching values across all properties
+   * Use the detected operator if input started with one (e.g., "!= test"), otherwise default to "="
+   */
   return {
     value: queryState.value,
-    options: generateFreeTextValueSuggestions(
+    options: createValueSuggestions(
       filteringOptions,
       queryState.operator ?? "=",
       queryState.value,
     ),
   };
+}
+
+/**
+ * Returns the valid operators for a given property.
+ * Currently returns all operators for all properties.
+ *
+ * TODO: Implement per-property operator configuration.
+ * This will allow properties to restrict which operators are valid.
+ * For example, a boolean property might only support "=" and "!=",
+ * while a text property supports ":", "!:", "^", etc.
+ *
+ * @returns Array of valid operators for the property
+ */
+function getValidOperatorsForProperty(
+  /* TODO: Will be implemented when each property can configure operator per instance */
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  _property: ParsedProperty,
+): QueryFilterOperator[] {
+  return QUERY_OPERATORS;
+}
+
+/**
+ * Filters the list of operators based on the provided prefix.
+ * If the prefix is empty, all operators are returned.
+ */
+function filterOperatorsByPrefix(
+  operators: QueryFilterOperator[],
+  prefix: string,
+): QueryFilterOperator[] {
+  if (!prefix) {
+    return operators;
+  }
+
+  return operators.filter((operator) => operator.startsWith(prefix));
+}
+
+function generatePropertySuggestions(
+  filteringProperties: ParsedProperty[] = [],
+  filterText = "",
+): OptionGroup<ParsedProperty>[] {
+  const filteredProperties: ParsedProperty[] = [];
+
+  for (const property of filteringProperties) {
+    if (!property) {
+      continue;
+    }
+
+    if (
+      matchesFilterText(
+        [
+          property.propertyLabel,
+          property.groupValuesLabel,
+          property.propertyGroup,
+        ].filter(Boolean),
+        filterText,
+      )
+    ) {
+      filteredProperties.push(property);
+    }
+  }
+
+  return createGroups(
+    filteredProperties,
+    (property) => property.propertyGroup,
+    "Properties",
+  );
+}
+
+function generateOperatorSuggestions(
+  property: ParsedProperty,
+  operatorPrefix = "",
+  filterText = "",
+): OptionGroup<AutoCompleteOption>[] {
+  const prefilteredOperators = filterOperatorsByPrefix(
+    getValidOperatorsForProperty(property),
+    operatorPrefix,
+  );
+
+  const operators: QueryFilterOperator[] = [];
+  for (const operator of prefilteredOperators) {
+    if (
+      matchesFilterText(
+        [operator, OPERATOR_LABELS[operator] ?? ""].filter(Boolean),
+        filterText,
+      )
+    ) {
+      operators.push(operator);
+    }
+  }
+
+  if (operators.length === 0) {
+    return [];
+  }
+
+  return [
+    {
+      label: "Operators",
+      options: operators.map((operator) => ({
+        value: buildQueryString(property.propertyLabel, operator, ""),
+        label: buildQueryString(property.propertyLabel, operator, ""),
+        description: OPERATOR_LABELS[operator] ?? "",
+      })),
+    },
+  ];
+}
+
+/**
+ * Creates value suggestions for autocomplete.
+ * When scopedProperty is provided, only shows values for that property (single group).
+ * When scopedProperty is omitted, searches across all properties (multiple groups).
+ */
+function createValueSuggestions(
+  filteringOptions: ParsedOption[] = [],
+  operator: QueryFilterOperator,
+  filterText = "",
+  scopedProperty?: ParsedProperty,
+): OptionGroup<AutoCompleteOption>[] {
+  const groups: Record<string, OptionGroup<AutoCompleteOption>> = {};
+
+  for (const option of filteringOptions) {
+    if (!option?.property) {
+      continue;
+    }
+
+    /* If scoped to a property, filter to only that property's options */
+    if (scopedProperty && option.property !== scopedProperty) {
+      continue;
+    }
+
+    /* Build search fields */
+    const searchFields = [
+      option.label,
+      ...(option.tags ?? []),
+      ...(option.filteringTags ?? []),
+    ];
+
+    if (!scopedProperty) {
+      searchFields.push(option.property.propertyLabel);
+    }
+
+    const matches = matchesFilterText(searchFields.filter(Boolean), filterText);
+
+    if (!matches) {
+      continue;
+    }
+
+    const groupLabel = option.property.groupValuesLabel || "Values";
+
+    if (!groups[groupLabel]) {
+      groups[groupLabel] = {
+        label: groupLabel,
+        options: [],
+      };
+    }
+
+    groups[groupLabel].options.push({
+      value: buildQueryString(
+        option.property.propertyLabel,
+        operator,
+        option.value,
+      ),
+      label: option.label,
+      tags: option.tags,
+      filteringTags: option.filteringTags,
+    });
+  }
+
+  return Object.values(groups).filter((group) => group.options.length > 0);
 }
 
 export { generateAutoCompleteOptions, QUERY_OPERATORS };
