+++ @navikt/core/react/src/data/token-filter/helpers/parse-query-text.ts
@@ -1,4 +1,10 @@
 import type { ParsedProperty, QueryFilterOperator } from "../TokenFilter.types";
+import { QUERY_OPERATORS } from "./generate-autocomplete-options";
+import {
+  matchFilteringProperty,
+  matchOperator,
+  matchOperatorPrefix,
+} from "./operators";
 
 type ParsedText =
   | {
@@ -76,73 +82,5 @@ function parseQueryText(
   };
 }
 
-/**
- * Operators ordered by specificity (longest/most specific first)
- * This ensures longer operators like ">=" and "<=" are matched
- * before shorter ones like ">" and "<"
- */
-const QUERY_OPERATORS: QueryFilterOperator[] = [
-  ">=",
-  "<=",
-  "!=",
-  "!:",
-  "!^",
-  "=",
-  ":",
-  "^",
-  ">",
-  "<",
-];
-
-/**
- * Match a property from the input text by longest property label.
- * Case-insensitive matching.
- */
-function matchFilteringProperty(
-  filteringProperties: ParsedProperty[],
-  text: string,
-): ParsedProperty | undefined {
-  const sortedProperties = [...filteringProperties].sort(
-    (a, b) => b.propertyLabel.length - a.propertyLabel.length,
-  );
-  return sortedProperties.find((prop) =>
-    text.toLowerCase().startsWith(prop.propertyLabel.toLowerCase()),
-  );
-}
-
-/**
- * Check if the input text is a valid prefix of any allowed operator.
- * Returns the prefix if valid, null otherwise.
- */
-function matchOperatorPrefix(
-  allowedOperators: QueryFilterOperator[],
-  filteringText: string,
-): string | null {
-  const trimmedText = filteringText.trim();
-
-  if (trimmedText.length === 0) {
-    return "";
-  }
-
-  const isValidPrefix = allowedOperators.some((operator) =>
-    operator.toLowerCase().startsWith(trimmedText.toLowerCase()),
-  );
-
-  return isValidPrefix ? trimmedText : null;
-}
-
-/**
- * Match an operator from the input text.
- * Operators are already sorted by specificity, so no re-sorting needed.
- */
-function matchOperator(
-  allowedOperators: QueryFilterOperator[],
-  text: string,
-): QueryFilterOperator | undefined {
-  return allowedOperators.find((operator) =>
-    text.toLowerCase().startsWith(operator.toLowerCase()),
-  );
-}
-
-export { QUERY_OPERATORS, parseQueryText };
+export { parseQueryText };
 export type { ParsedText };
